Fix passenger postinst to use correct command
The passenger postinst was using the output of `puppetca --configprint`, which
on 2.7.x produces man-page like output, which made the cert generation command
generate a cert for each of the words in the manpage until it hit a duplucate
and errored out. This patch replaces the `puppetca --configprint` with `puppet
master --configprint`, so that the output is what is expected. The `puppetca
--generate` is replaced with `puppet ca --generate` to match the 2.7.x syntax.
It also replaces other calls to `puppet --configprint` with `puppet master
--configprint` so it is clear where the configuration is coming from.