(#11509, #11656) Fix puppetmaster-passenger postinst
The postinst script was calling the incorrect puppet command and the apache
template referencing an invalid docroot. This patch updates the puppet command
to use `puppet cert generate` instead of `puppet ca generate` so that the
ca_crl.pem gets generated. It also seds the apache config to point to the
new/valid docroot in /usr/share/puppet/rack/puppetmasterd so that rack can
successfully start.