(#12186) fix puppet daemon status in hardy
In Ubuntu Hardy, the /lib/lsb/init-functions script for managing
services does not implement the status_of_proc() function, which
is a standard method of getting proc info for service status info
in init scripts.  The status_of_proc() function was implemented in
later versions of lsb.  Currently, the init scripts in the
community-debian release of puppet source the init-functions script
and call to status_of_proc() for status info, which fails in hardy
bacause the status_of_proc() function is not implemented.

This patch modifies the puppet.init and puppetmaster.init scripts
in the puppet debian package to implement a version of status_of_proc
(from later versions of lsb). Rather than modify
/lib/lsb/init-functions directly, this patch adds a define-if section
to the status function of both init scripts that defines
status_of_proc() if it is not found. This allows for future/modified
versions of status_of_proc in init-functions to be honored and
should not affect any version post-Hardy.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>